Tercera práctica integradora

+++ Proyecto/src/controllers/api/products.controller.js
@@ -2,6 +2,12 @@ import CustomError from '../../services/errors/CustomError.js';
 import EErrors from '../../services/errors/enum.js';
 import { generateProductErrorInfo } from '../../services/errors/info.js';
 import * as productsService from '../../services/products.service.js'; 
+import { ResultNotFound } from '../../utils/customExceptions.js';
+import jwt from 'jsonwebtoken';
+import config from '../../config/config.js';
+import { RolForbiden } from '../../utils/customExceptions.js';
+
+const PRIVATE_KEY = config.secret;
 
 const getProductsPaginate = async (req, res) => {  
     try{
@@ -13,7 +19,7 @@ const getProductsPaginate = async (req, res) => {
         req.logger.info(`Solicitud procesada: ${req.method} ${req.url}`);
     } catch(error){
         res.status(500).send({error});
-        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()}`);
+        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()} - ${error}`);
     }
 
 };
@@ -23,24 +29,27 @@ const getProduct = async(req,res)=> {
         const pid = req.params.pid;
 
         const result = await productsService.getProduct(pid);
-        
-        result ? res.send({ status: 'success', result }) : res.status(400).send({status: 'error', message:`Producto no encontrado`})
+        // result ? res.send({ status: 'success', result }) : res.status(400).send({status: 'error', message:`Producto no encontrado`})
+        res.sendSuccess(result);
         req.logger.info(`Solicitud procesada: ${req.method} ${req.url}`);
     } catch (error) {
-        res.status(500).send({ status:"error", message: error});  
-        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()}`);
+        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()} - ${error}`);
+        if(error instanceof ResultNotFound){
+            res.sendClientError('Producto no encontrado')
+        }
+        res.sendServerError(error);
     }
     
 };
 
 const saveProduct = async (req, res) => {
     try{
-        const { title,description,code,price,status,stock,category,thumbnails,id } = req.body;    
+        const { title,description,code,price,status,stock,category,thumbnails,id, owner } = req.body;    
 
         if(!title || !description || !code || !price || !stock || !category){
             // return res.status(400).send({status: 'error', message:'Valores incompletos'});
             throw CustomError.createError({
-                name: 'UserError',
+                name: 'ProductError',
                 cause: generateProductErrorInfo({
                     title,
                     description,
@@ -54,6 +63,9 @@ const saveProduct = async (req, res) => {
             });
         } 
 
+        const token = req.cookies.token;
+        const decodedToken = jwt.verify(token, PRIVATE_KEY);
+
         const product = {
             title,
             description,
@@ -63,7 +75,8 @@ const saveProduct = async (req, res) => {
             stock, 
             category,
             thumbnails,
-            id
+            id,
+            owner: decodedToken.rol==="PREMIUM" ? decodedToken.email : "ADMIN"
         };
         
         const result = await productsService.saveProduct(product);
@@ -71,7 +84,7 @@ const saveProduct = async (req, res) => {
         res.send({ result: 'success', result});
     }catch(error){
         res.status(500).send({error});
-        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()}`);
+        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()} - ${error}`);
     }
 
     // const product = req.body;
@@ -86,13 +99,37 @@ const updateProduct = async (req,res)=> {
         const productReq = req.body;
         const pid = req.params.pid;
 
-        const result = await productsService.updateProduct(productReq, pid);
+        const token = req.cookies.token;
+        const decodedToken = jwt.verify(token, PRIVATE_KEY);
 
-        result ? res.send({status: 'sucess', message:'Producto Modificado'}) : res.send({status: 'error', message:`Producto no encontrado`})
-        req.logger.info(`Solicitud procesada: ${req.method} ${req.url}`);
+        const product = await productsService.getProduct(pid);
+
+
+        if(decodedToken.rol === 'ADMIN') {
+            const result = await productsService.updateProduct(productReq, pid);
+
+            req.logger.info(`Solicitud procesada: ${req.method} ${req.url}`);
+            return res.send({status: 'sucess', message:'Producto Modificado'});
+        } 
+
+        if(decodedToken.rol === 'PREMIUM' && decodedToken.email === product.owner) {
+            const result = await productsService.updateProduct(productReq, pid);
+            
+            req.logger.info(`Solicitud procesada: ${req.method} ${req.url}`);
+            return res.send({status: 'sucess', message:'Producto Modificado'});
+        } else {
+            throw new RolForbiden('User does not owner of product');
+        }
+        
     }catch(error){
+        if(error instanceof ResultNotFound){
+            res.sendClientError('Carrito no encontrado')
+        }
+        if(error instanceof RolForbiden){
+            res.sendClientError('User does not owner of product')
+        }
         res.status(500).send({status: 'error'});
-        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()}`);
+        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()} - ${error}`);
     }
     // await manager.update(productReq, pid);
 };
@@ -101,13 +138,37 @@ const deleteProduct =  async (req,res)=> {
     try{
         const pid = req.params.pid;
 
-        const result = await productsService.deleteProduct(pid)
+        const token = req.cookies.token;
+        const decodedToken = jwt.verify(token, PRIVATE_KEY);
 
-        result ?  res.send({status: 'sucess', message:'Producto eliminado'}) : res.send({status: 'error', message: 'Producto no encontrado'})
-        req.logger.info(`Solicitud procesada: ${req.method} ${req.url}`);
+        const product = await productsService.getProduct(pid);
+
+
+        if(decodedToken.rol === 'ADMIN') {
+            const result = await productsService.deleteProduct(pid)
+
+            req.logger.info(`Solicitud procesada: ${req.method} ${req.url}`);
+            return res.send({status: 'sucess', message:'Producto Eliminado'});
+        } 
+
+        if(decodedToken.rol === 'PREMIUM' && decodedToken.email === product.owner) {
+            const result = await productsService.deleteProduct(pid)
+
+            req.logger.info(`Solicitud procesada: ${req.method} ${req.url}`);
+            return res.send({status: 'sucess', message:'Producto Eliminado'});
+        } else {
+            throw new RolForbiden('User does not owner of product');
+        }
+
     }catch(error){
+        if(error instanceof ResultNotFound){
+            res.sendClientError('Carrito no encontrado')
+        }
+        if(error instanceof RolForbiden){
+            res.sendClientError('User does not owner of product')
+        }
         res.status(500).send({status: 'error'});
-        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()}`);
+        req.logger.error(`${req.method} en ${req.url} - ${new Date().toISOString()} - ${error}`);
     }
 
     // const removedProduct = await manager.deleteById(pid)
